getDateAndDigestAndSize(): handle creation time not being set
The image creation timestamp that we get back as part of the result
structure from github.com/containers/image/types/Image.Inspect() can be
unset (i.e., nil), since it's optional in the OCI image spec.  This
means we should check that the creation timestamp isn't a nil pointer
before attempting to dereference it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>