manifest push --format: force an image type, not a list type
The image library's image copying API expects the type of a single image
in the options structure's ForceManifestMIMEType field, and it infers
the corresponding list type when it's copying multiple images.

Previously, though, we were passing in a list type, which was also being
used as the target type for converting single images, which would fail.
Switch to passing in the single-image type instead, and infer the single
type in our internal API if a consumer passes us a list type.

This makes it easier to notice that we force v2s1 images to be converted
to v2s2 images when we push them as part of a list, even if the registry
doesn't reject v2s1 images.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>