Improve remote manifest retrieval
Before this patch, it was not possible to retrieve a remote manifest
just by specifying the image name, like:

```
> buildah manifest inspect alpine
Invalid image name "alpine", expected colon-separated transport:reference
```

It was possible to get the manifest via:

```
> buildah manifest inspect docker://alpine
…
```

But after pulling the image into the local storage, this works not any
more:

```
> buildah pull alpine
e7d92cdc71feacf90708cb59182d0df1b911f8ae022d29e8e95d75ca6a99776a
> buildah manifest inspect docker://alpine
manifest from image … is of type
"application/vnd.docker.distribution.manifest.v2+json", which
is not a list type
ERRO exit status 1
```

This means we now collect a list of possible local or remote manifests
and try to resolve them sequentially. This enables us to fallback to the
remote location if the locally fetched manifest is not an actual
manifest. It also enables us to see the remote manifest via:

```
> ./buildah manifest inspect alpine
{
    …
}
```

Signed-off-by: Sascha Grunert <sgrunert@suse.com>