Add Containerfile to build a versioned stable image on quay.io
I've been using this Containerfile to build versioned stable images
of Buildah for quay.io.  I build these images if I miss grabbing the
'latest' image before the next version goes to latest, or if we've
made changes to how the image is built and I need to redo the older
versions.

As we've had a few updates by different people to the Containerfiles
in these directories, I thought I'd add this one too so it can be
updated by others if needed.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>