image with dup layers: we now have one on quay
...and a process to create future ones. Thanks to Nalin for
a short patch to skopeo that allows pushing images with dups.

This commit updates make-v2sN to add "with-dups" functionality;
it simply creates a v2s1 image, then uses jq (JSON tool) to
duplicate the first .fsLayers and .history elements. Updated
instructions so others can (we hope) do this in the future
should it ever be necessary (we hope not).

Update the checks that confirm that the image we pull has
dup layers - this is a guarantee that we're testing what
we think we're testing.

Signed-off-by: Ed Santiago <santiago@redhat.com>