Make pull-always always pull
I've changed the vendor piece of c/storage that will need to change
with this commit.  If @mtrmac is agreeable with this concept, I'll
work up a proper PR for c/storage.

The motivation for this changes is we currently say in our docs that
`--pull-always` always pulls from the registry regardless of what's
locally stored.

That turns out to NOT be the case, we do reuse some of the images/layers
if they are present locally.

This change passes the `--pull-always` flag down to the c/storage code
so it can decide what to do.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>