Make image history work correctly with new args handling
Previously, every build-arg was recorded in the image history
even if the arg was not accessible to that layer when the build
was run.

This commit fixes that by ensuring that args are only added to
history when they are in scope. This is tracked in the
imagebuilder.Builder struct which is now accessible through
the stage reference in the StageExecutor

Fixes #2210

Signed-off-by: Nick Carboni <ncarboni@redhat.com>