Fix bud-build-arg-cache test
This test had a false assumption causing it to fail with my new
changes to ARG handling.

When a build arg is provided a layer should only be re-built if
that build arg could change the output of that layer. For this
reason the previous Dockerfile would (correctly) be pulled from
cache even if a build arg was provided. To force a rebuild, we need
a RUN command after the ARG command that we're changing.

Signed-off-by: Nick Carboni <ncarboni@redhat.com>