dockerignore tests : remove symlinks, rework
rpm 4.16 alpha in rawhide has broken our koji build:

   https://github.com/containers/buildah/issues/2264

   https://github.com/rpm-software-management/rpm/issues/1159

TL;DR we can no longer package the 'circular-symlink' file
used in tests. This PR removes that and other symlinks from
the git working tree, and reworks the test so it runs from
a temporary directory and creates all necessary symlinks.

Also, since the test as written was simply checking "did it
run" without actually testing anything, add a test to confirm
that the files added to the image are exactly what we expect.

Signed-off-by: Ed Santiago <santiago@redhat.com>