Downgrade siruspen/logrus from 1.4.2
The upgrade of logrus forces to update the dependency in go module
consumers as well.

The new locking in logrus 1.5.0 introduces regressions in conjunction to
deadlocks when using custom hooks, which is the case in CRI-O for
example.
Having it point to v1.5.0 here is forcing cri-o to upgrade it as well
when we update c/buildah to v1.14.6.

Let's downgrade logrus till sirupsen/logrus#1122
is fixed.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>