Fix (make vendor)
(export a=b command args) does not run (command args) with a=b,
it sets $a to b, and marks variables $a $command $args as exported,
i.e. (command args) is not run.

So, before https://github.com/containers/buildah/pull/2287 we were not actually
running (go mod tidy), and now we are not running (go mod vendor).

Fix that.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>