Add containers.conf to stablebyhand build
Add the containers.conf in the build process of the
"stablebyhand" Containerfile which is used to build
a particular variant of Buildah by hand for use in quay.io.

The change for that is the ADD of the container.conf in
GitHub.com and emulates @rhatdan 's work in #2284

Also bumped the version to a more recent release while
I was in here.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>