Add VFS additional image store to container
For the overlay driver, the buildah container was given an additional
image store. Unfortunatley, overlay doesn't work out-of-the-box in a
Kubernetes cluster do to the lack of appropriate access to /dev/fuse.
In Minikube, it's non-trivial to provide that access, but it is trivial
to use the VFS storage driver.

Previously, the VFS storage driver would fail due to not having the
appropriate directory structure and required files in the additional
image store. This commit adds those files to the pre-built container
images to allow for easy use in Kubernetes.

Signed-off-by: Zvi "CtrlZvi" Effron <viz_skywalker+GitHub@outlook.com>