imagebuildah: wait for stages that might not have even started yet
When checking if a name specified for a FROM instruction, or a source
for a COPY instruction, was a stage for which it needed to wait,
StageExecutor was checking the Executor's stages list, to which stages
are added only as they're started.  This could cause it to mistakenly
write off a name if the corresponding stage hadn't yet been started, and
assume that the named stage was actually an image.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>