bud.bats: correct .dockerignore integration tests
The "bud with .dockerignore" test shouldn't be able to copy "subdir"
from the build context to the build container when "subdir" is the next
to last item listed in the build context's .dockerignore file, even
though the "!*/sub1* after it might suggest otherwise, for
compatibility.  Add a second Dockerfile which includes every step except
the last one, which we now expect to fail, to ensure that we succeed up
until that point.

The "bud with .dockerignore - 3" test shouldn't expect to be able to
copy "test1.txt" from the build context, when it's the last item
mentioned in the build context's .dockerignore file.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>