Add buildah.IsContainer interface
This interface will allow callers to specify a storage container id
and the store and return whether or not the container is a buildah
container.

We want to add the ability to for Podman to identify containers in storage
that it did not create or were created via podman build command.

Podman will use this information to help user identify where the container
came from.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>