Update Buildah Tutorial to address BZ1867426
A bugzilla was created due to a misunderstaning gleened from the
Buildah tutorial.  The person runing the tutorial followed it exactly,
but the example blew up on RHEL for them.  The reason being is the
`--releasever` option has to use a valid version for the host's OS
when it's working on a scratch container or dnf will fail.

This adds clarifiction to the document to make that clear.

https://bugzilla.redhat.com/show_bug.cgi?id=1867426

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>