contrib/cirrus/lib.sh: don't use CN for the hostname
When generating a certificate to be used by a registry server that we're
bringing up, instead of setting a CN value in the subject name to
"localhost" to pass the name check that a client makes, use a subject
alternative name extension.

The "compare the hostname we're given with a CN from the subject name
field" method stops working if the client is built with Go 1.15.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>