Fix buildah mount to display container names not ids
Currently we are displaying containerid rather then containername,
which is unknown to users.  Users refer to the container names all
of the time, so we should show the mounted containers based on names.

This change eliminates the need for the truncate flag on buildah mount
as well, so this PR hides it.

Fix up the man page documentation to use container names and show
proper output.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>
Co-authored-by: Sascha Grunert <sgrunert@suse.com>
Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>