Add(): fix handling of relative paths with no ContextDir
Add() is supposed to handle relative paths when the ContextDir value
passed to it is not set, but it hasn't been doing that correctly since
it was overhauled.  Correct it to recognize relative paths at the
function start, when ContextDir is not set, by combining them with the
current directory to convert them to absolute paths.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>