New "add relative" test: make it work in gating env
Gating tests are failing because the new add.bats:relative
test assumes it is cd'ed to a known location. This is not
true in gating tests (neither Fedora nor RHEL).

Fix: for this test, 'cd ${TESTSDIR}'. I've confirmed on my
end that this affects only the scope of the current test,
but I don't actually know if that's a BATS guarantee.

Signed-off-by: Ed Santiago <santiago@redhat.com>