CI: run gating tasks with a lot more memory
Try running the gating test with a lot more memory (12Gb, just shy of
the maximum of 13 allowed for our default 2 vcpu configuration, instead
of the default of 4Gb that we use elsewhere).  We're seeing the linting
process getting killed frequently without any additional explanation,
and our current best guess is that it's the OOM killer.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>