imagebuildah: cache should take image format into account
When evaluating cache candidates during a build, only consider an image
if it's in the same format that we're attempting to build.  That way, we
won't mistakenly try to use an OCI format image as a cache when we're
attempting to build an image in Docker format because we're using
configuration features specific to that format, but we forgot to specify
the format during a previous attempt.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>