imagebuildah: change when we decide to create layers
Change our strategy for creating layers from:
 * Always create layers for ADD, COPY, and RUN
to:
 * Create layers for ADD, COPY, RUN, and WORKDIR if
   they made changes to the filesystem.  This includes
   creating the WORKDIR directory if it doesn't already
   exist.

It's often a subtle difference, but it fixes cases where a WORKDIR
instruction wasn't followed by RUN, so we didn't end up accidentally
creating it ourselves.

We're not perfect here, because our RUN creates targets for bind mounts
that `docker build` avoids, but we're closer than we were before.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>