imagebuildah: use overlay for volumes when using overlay
When we're using the overlay driver (which means we know overlay is
available), use it to make volumes appear to be writeable during RUN
instructions instead of saving/restoring their contents.

This avoids having to copy the contents of the volume directory before
each RUN instruction, and having to remove and extract the contents
after each RUN instruction, which should be faster, particularly if the
amount of content in that volume location is large.

For empty directories, it will at least avoid adding an "opaque"
notation for the directory in a layer that might otherwise be empty.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>