Compute diffID for mapped-layer at creating image source
If UID and GID mappings are specified, the container has a
mapped-layer, whose diffID is not computed when created.
Committing the image fails due to lack of diffID. This fix
computes diffID at creating an image source if a layer
doesn't have a diffID (UncompressedDigest).

This fix also tests if a container with UID and GID mappings
can be committed.

Signed-off-by: Hironori Shiina <shiina.hironori@fujitsu.com>