intermediateImageExists: ignore images whose history we can't read
When attempting to read the history of an image in order to try to check
if it's suitable as a cache entry for an image we're building, if we
fail to read its history, ignore the image instead of failing.

If the image was pulled into local storage for a different architecture,
or it's a list with no entry in it for the local architecture, or if
it's a list which references an image for the local architecture that we
haven't pulled down, we'll get an error back from the image library, and
we don't want to fail because of that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>