Not prefer use local image create/add manifest
Avoid first checking the image from local storage for `manifest create` and `manifest add`
since the local image does not include other entries of the list from the registry.
`--all` flag of `manifest create` and `manifest add` can not add all of the lists as expected.
So first attempt to get manifest from the registry and if not successful, fall back to use local images.

Signed-off-by: Qi Wang <qiwan@redhat.com>