Cirrus: Support new VM Images in get_ci_vm.sh
Due to necessary changes in both scripts and within images, updates are
necessary for this script to function.  These changes also bring
the script very close to being identical to the same in the
containers/podman repository.  The intent is to support a future
where this script can simply be shared across repos.

Signed-off-by: Chris Evich <cevich@redhat.com>