copier.PutOptions: add an "IgnoreDevices" flag
Prior to 1.16, when ADDing contents to a working container, if we were
being run by an unprivileged user using a user namespace, content that
was a device node would be ignored.

Add a flag in copier.PutOptions that tells copier.Put() to ignore
entries that are either a device, or a hard link to a device.

Make buildah.Add() set the IgnoreDevices flag in PutOptions when
libcontainer says we're running in a user namespace.

Together, these two changes should restore the earlier behavior.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>