Cirrus: Clarify task names / improve readability
This change mirrors a similar change made in early 2020 for the
containers/podman repository.  Primarily, it alters the appearance of
the task names in the Github WebUI to be more readable by humans.
Also, the former "gate" task is renamed to "smoke".  This prevents
confusion with downstream Fedora/RHEL post-packaging testing which
is(also referred to as the "gate" tests.

Finally, task-dependencies are altered to support a cleaner flow which
will be implemented in a future commit.

Signed-off-by: Chris Evich <cevich@redhat.com>