Fix config-flags-verification test on F33
On newer versions of `time`, the error message quotes the bad value.
This leads to an erroneous failure in the test:

```
working-container
time: invalid duration ABCD'
time: invalid duration "ABCD"'
```

The comparison used for `expect_output --substrig` supports basic
regular expressions.  Encode the expected value in the test to be
tolerant of quoting in the output value.

Signed-off-by: Chris Evich <cevich@redhat.com>