Fix conformance test false-failures
Prior to Docker 20.10, failed COPY/ADD instructions result in error
messages containing the string `no such file or directory`.  This was
changed in docker to a string containing `file not found`.  Since
the conformance tests potentially need to work with multiple docker
versions, they need to be sensitive to both errors.

Signed-off-by: Chris Evich <cevich@redhat.com>