WIP: Add local name to short names list
This is very much a WIP.  Part of this needs to go in c/image, the
rest in Buildah.  But I wanted to put this up for a POC review.

In short, when we're building an image, we see if the image is
known locally.  If it is, and the user gives us the ID for it,
we accept it and continue on with the build.

However, if the user gave us the name, especially in a
`FROM` statement in a Containerfile, we first found that it
was locally, then resolve the image name using all of
the registries that we know, but neglect to add the local
image to that list.

Now if we find the image locally, this code will pass that
name down to be added to the top of the list that we present
to the user.

Once this approach is OK'd, I'll put a PR together
in c/image and vendor into Buildah, that is unless
@vrothberg beats me to the punch.

Resolves: #2904

Test run:

```
STEP 1: FROM alpine
STEP 2: RUN touch /foo
STEP 3: ONBUILD RUN touch /bar
STEP 4: COMMIT onbuild-image
Getting image source signatures
Copying blob 777b2c648970 skipped: already exists
Copying blob c440ad84933e done
Copying config 0f963b3733 done
Writing manifest to image destination
Storing signatures
--> 0f963b37335
0f963b373356dc17643f3b073a77cc8be37b401837bdf9390d196a4915196f98
STEP 1: FROM onbuild-image
? Please select an image:
  ▸ onbuild-image:latest
    docker.io/library/onbuild-image:latest
    registry.fedoraproject.org/onbuild-image:latest
    registry.access.redhat.com/onbuild-image:latest
↓   registry.centos.org/onbuild-image:latest

```

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>