Cherry pick localhost fix and update CI configuration for release-1.19 branch
Cherry pick @vrothberg's "use local image name for pull policy checks" #2908
and update the cirrus and git validations so the test will run on this new(ish)
branch.  From @vrothberg:

Some pull policies require to first look up a local image and compare
that to the remote counter part. When looking up the remote image, we
need to make sure to use the name of the local image, if it exists.

This fixes a bug where a short name resolved to an image with the
"localhost/" prefix. This prefix is only used for local image look ups
via shortnames.ResolveLocally. Hence, when looking up the remote
counter part, we must preserve this prefix.

Fixes: #2904

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>