blobcache: fix sequencing error
Wait for the goroutine that's saving a decompressed copy of the blob to
finish before checking the digest that it's computing, to avoid a race
condition which would cause us to read the digest value before we're
finished computing it.
Refuse to get involved when compression isn't gzip, pending a bigger
update to handle zstd correctly.
Set CompressionOperation and MediaType when suggesting substitutions.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>