Stop testing directory permissions with latest docker
Current version of Docker, has a bug we believe, that is
creating top level directories when using ADD and COPY of
tar balls. Basically the directories end up with a 777
permissions, Buildah creates these with 755, which we
believe is correct, and matches older versions of Docker.

We need to revert this patch once we have a version of Docker
that creates these direcories with the corret permissions.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>