'make validate': require PRs to include tests
This is an almost-clone of podman PR 8911[1], in which Ed
ticks off developers by requiring that PRs include tests.

 [1] https://github.com/containers/podman/pull/8911

Motivated by the just-merged #2973, which is exactly the sort
of tiny little change that can have huge consequences down
the line and/or be easy to undo.

Changes from the podman version: allow '[NO NEW TESTS NEEDED]'
in addition to just '[NO TESTS NEEDED]', and make the former
the default suggestion. It's clearer. And, of course, update
the test-suite cases to reflect the buildah git tree.

Unrelated parasitic change: I removed the '@'s from 'validate'
steps; this will make it easier to (1) confirm that all desired
rules were invoked, and (2) on failure, see the exact command
that failed, making it quicker/easier for dev to reproduce.

Signed-off-by: Ed Santiago <santiago@redhat.com>