Add more support for removing config
Currently we can remove configuration data in buildah config
by adding a trailing "-". This PR adds support for this in --port
calls.  Also added support for clearing all config for a specified
option, if the user specifies "-".

Currently the code blocks setting ANNOTATIONS and LABELS without a
value.  This is broken and should be allowed.

Similarly we were not expaning envioronment variables from the host when
they were not set.

podman run --env foobar
is valid,
so
buildah config --env foobar
should also be valid.

Fixes: https://github.com/containers/buildah/issues/2859

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>