pr-should-include-tests: recognized "renamed" tests
git tries to recognize renamed files. This isn't always
as helpful as intended. Turn it off, so we'll always see
files as 'A'dded.

This is a simple copy of podman PR 9468, but without a
test case because I can find no instances of a .bats
file being renamed in the buildah git history.

Signed-off-by: Ed Santiago <santiago@redhat.com>