Check local image to match system context
Currently if you pull one image and then pullifmissing with
a different --arch, Buildah does not pull the different arch,
even though the arch is missing.

This PR checks the existing image to see if it matches the arch, os
and variant of the specified image, before using the local image.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>