tests: prefetch: use buildah, not podman, for pulls
The _prefetch helper has proven invaluable in reducing pull-
related flakes. However, it is now failing because it uses
podman to pull, and buildah tests rely on shortnames, and
podman no longer works with shortnames.

Solution: use buildah, not podman, for initial image pulls.
(The reason we used podman in the first place is because
buildah does not have save/load. But since image store
is shared, we can use buildah for pull and podman for save
and load.)

Add shortname config to registries.conf. I don't think it's
actually necessary or even used, but we can tweak it later.

And, fix one test in from.bats that was using inconsistent
names between buildah and podman.

Signed-off-by: Ed Santiago <santiago@redhat.com>